htlcswitch/test: move test code
